python-flask-version/requirements.txt
flask
firebase_admin
google-api-python-client
google-cloud-secret-manager
werkzeug
waitress